+++ app/Livewire/Harga/HargaDelete.php
@@ -3,6 +3,7 @@
 namespace App\Livewire\Harga;
 
 use App\Models\Harga;
+use App\Models\Pesanan;
 use Livewire\Component;
 use Livewire\Attributes\On;
 use Livewire\Attributes\Locked;
@@ -29,8 +30,17 @@ class HargaDelete extends Component
 
     public function del()
     {
+        $harga = Pesanan::where('id_harga', $this->id)->get();
+
+        if ($harga->isNotEmpty()) {
+            $this->dispatch('notify', title: 'fail', message: 'Data gagal dihapus, harga sedang digunakan');
+            $this->modalHargaDelete = false;
+            return;
+        }
+        
         $del = Harga::destroy($this->id);
 
+        // dd($del);
         $del
             ? $this->dispatch('notify', title: 'success', message: 'Data berhasil dihapus')
             : $this->dispatch('notify', title: 'fail', message: 'Data gagal dihapus');
